Separate type definitions
Separate the functionality from type definitions. This makes the script
more readable, but makes no functionality changes
